unused packages were uninstalled and some cleaning and made routing simpler
