+++ package.json
@@ -11,13 +11,10 @@
   },
   "dependencies": {
     "@emotion/react": "^11.11.4",
-    "allotment": "^1.20.0",
     "react": "^18.2.0",
     "react-awesome-reveal": "^4.2.8",
-    "react-compare-slider": "^3.1.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.3",
-    "react-swipeable": "^7.0.1"
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
