--- src/components/pages/FAQs/FAQs.css
@@ -1,28 +1,0 @@
-:root {
-    --image-comparison-slider-width: 100vw;
-    --image-comparison-slider-handle-width: 50px;
-}
-
-#image-comparison-slider {
-    width: var(--image-comparison-slider-width);
-
-}
-
-#image-comparison-slider .handle {
-    left: calc(50% - var(--image-comparison-slider-handle-width)/2);
-    width: var(--image-comparison-slider-handle-width);
-    /* transition: 0.25s cubic-bezier(.03,.98,.52,.99); */
-}
-
-#image-comparison-slider .handle-circle {
-    width: var(--image-comparison-slider-handle-width);
-    height: var(--image-comparison-slider-handle-width);
-}
-
-#image-comparison-slider .slide {
-    width: var(--image-comparison-slider-width);
-}
-
-#image-comparison-slider .img-wrapper {
-transform: translateZ(0);
-}
